+++ ChangeLog.md
@@ -15,6 +15,7 @@
     - Prevent a form to be submitted when hitting Enter during edition
     - Make sure we have a createdAt/createdBy data after schema modification
     - Provide a full publication of the contents collection
+    - Provide a workaround for Meteor bug #12524 when font files are to be loaded
  
 ### 1.3.0